+++ lib/src/forms/print-percentile.dart
@@ -88,7 +88,7 @@ class PrintPercentile extends BasePrint
     double xo = xorg;
     double yo = yorg;
     var data = src.calc;
-    lineWidth = cml(0.03);
+    lineWidth = cm(0.03);
 
     titleInfo = titleInfoBegEnd(src);
     List<EntryData> times = data.entries;
@@ -122,11 +122,11 @@ class PrintPercentile extends BasePrint
     {
       vertCvs.add({
         "type": "line",
-        "x1": cmx(i * colWidth),
-        "y1": 0,
-        "x2": cmx(i * colWidth),
-        "y2": cmy(lineHeight * gridLines + 0.25),
-        "lineWidth": cml(lw),
+        "x1": cm(i * colWidth),
+        "y1": cm(0),
+        "x2": cm(i * colWidth),
+        "y2": cm(lineHeight * gridLines + 0.25),
+        "lineWidth": cm(lw),
         "lineColor": i > 0 && i < 24 ? lc : lcFrame
       });
       if (i < 24)horzStack.add({
@@ -140,11 +140,11 @@ class PrintPercentile extends BasePrint
     {
       horzCvs.add({
         "type": "line",
-        "x1": cmx(-0.2),
-        "y1": cmy((gridLines - i) * lineHeight - lw / 2),
-        "x2": cmx(24 * colWidth + 0.2),
-        "y2": cmy((gridLines - i) * lineHeight - lw / 2),
-        "lineWidth": cml(lw),
+        "x1": cm(-0.2),
+        "y1": cm((gridLines - i) * lineHeight - lw / 2),
+        "x2": cm(24 * colWidth + 0.2),
+        "y2": cm((gridLines - i) * lineHeight - lw / 2),
+        "lineWidth": cm(lw),
         "lineColor": i > 0 ? lc : lcFrame
       });
       if (i > 0)
@@ -174,36 +174,36 @@ class PrintPercentile extends BasePrint
       "canvas": [
         {
           "type": "rect",
-          "x": cmx(0.0),
-          "y": cmy(yHigh),
-          "w": cml(24 * colWidth),
-          "h": cml(yLow - yHigh),
+          "x": cm(0.0),
+          "y": cm(yHigh),
+          "w": cm(24 * colWidth),
+          "h": cm(yLow - yHigh),
           "color": "#0f0",
           "fillOpacity": 0.5
         },
         {
           "type": "line",
-          "x1": cmx(0.0),
-          "y1": cmy(yHigh),
-          "x2": cmx(24 * colWidth),
-          "y2": cmy(yHigh),
-          "lineWidth": cml(lw),
+          "x1": cm(0.0),
+          "y1": cm(yHigh),
+          "x2": cm(24 * colWidth),
+          "y2": cm(yHigh),
+          "lineWidth": cm(lw),
           "lineColor": "#0a0"
         },
         {
           "type": "line",
-          "x1": cmx(0.0),
-          "y1": cmy(yLow),
-          "x2": cmx(24 * colWidth),
-          "y2": cmy(yLow),
-          "lineWidth": cml(lw),
+          "x1": cm(0.0),
+          "y1": cm(yLow),
+          "x2": cm(24 * colWidth),
+          "y2": cm(yLow),
+          "lineWidth": cm(lw),
           "lineColor": "#0a0"
         },
         {"type": "rect", "x": 0, "y": 0, "w": 0, "h": 0, "color": "#000", "fillOpacity": 1}
       ]
     };
     var percGraph = {"absolutePosition": {"x": cmx(xo), "y": cmy(yo)}, "canvas": []};
-    var percLegend = LegendData(cmx(xo), cmy(yo + lineHeight * gridLines + 1.0), cmx(8.0), 100);
+    var percLegend = LegendData(cmx(xo), cmy(yo + lineHeight * gridLines + 1.0), cm(8.0), 100);
 
     if (addPercentileGraph(percGraph, percList, 10, 90, "#aaf"))addLegendEntry(percLegend, "#aaf", msgPercentile1090);
     if (addPercentileGraph(percGraph, percList, 25, 75, "#88f"))addLegendEntry(percLegend, "#88f", msgPercentile2575);
@@ -231,14 +231,14 @@ class PrintPercentile extends BasePrint
     {
       if (entry.percentile(high) != entry.percentile(low))ret = true;
       x = glucX(entry.time);
-      ptsHigh.add({"x": cmx(x), "y": cmy(glucY(entry.percentile(high)))});
-      if (high != low)ptsLow.insert(0, {"x": cmx(x), "y": cmy(glucY(entry.percentile(low)))});
+      ptsHigh.add({"x": cm(x), "y": cm(glucY(entry.percentile(high)))});
+      if (high != low)ptsLow.insert(0, {"x": cm(x), "y": cm(glucY(entry.percentile(low)))});
 //        ptsHigh.add({"x": cmx(x), "y": cmy(glucY(entry.median))});
     }
     x = glucX(DateTime(0, 1, 1, 23, 59, 59));
-    ptsHigh.add({"x": cmx(x), "y": cmy(glucY(percList.values.first.percentile(high)))});
-    if (high != low)ptsLow.insert(0, {"x": cmx(x), "y": cmy(glucY(percList.values.first.percentile(low)))});
-    var area = {"type": "polyline", "lineWidth": cml(lw), "closePath": high != low, "fillOpacity": 0.5, "points": []};
+    ptsHigh.add({"x": cm(x), "y": cm(glucY(percList.values.first.percentile(high)))});
+    if (high != low)ptsLow.insert(0, {"x": cm(x), "y": cm(glucY(percList.values.first.percentile(low)))});
+    var area = {"type": "polyline", "lineWidth": cm(lw), "closePath": high != low, "fillOpacity": 0.5, "points": []};
     (area["points"] as List).addAll(ptsHigh);
     if (high != low)
     {
@@ -249,9 +249,9 @@ class PrintPercentile extends BasePrint
     (percGraph["canvas"] as List).add(
       {"type": "rect", "x": 0, "y": 0, "w": 0, "h": 0, "color": "#000", "fillOpacity": 1});
     (percGraph["canvas"] as List).add(
-      {"type": "polyline", "lineWidth": cml(lw), "closePath": false, "lineColor": color, "points": ptsHigh});
+      {"type": "polyline", "lineWidth": cm(lw), "closePath": false, "lineColor": color, "points": ptsHigh});
     (percGraph["canvas"] as List).add(
-      {"type": "polyline", "lineWidth": cml(lw), "closePath": false, "lineColor": color, "points": ptsLow});
+      {"type": "polyline", "lineWidth": cm(lw), "closePath": false, "lineColor": color, "points": ptsLow});
 
     return ret;
   }